refactor(daengle): 폴더구조 수정 및 style 파일 통합
