+++ apps/daengle/src/interfaces/auth/index.ts
@@ -15,3 +15,13 @@ export interface PetInfoFormType {
   breed: string;
   petWeight: 'SMALL' | 'MEDIUM' | 'LARGE';
 }
+
+export interface UserProfileInfoEditForm {
+  image: File | null;
+  nickname: string;
+  isAvailableNickname: boolean;
+}
+export interface UserProfileInfoEditFormType {
+  form: UserProfileInfoEditForm;
+  isAvailableNickname: boolean;
+}
