--- apps/daengle/src/pages/mypage/user-profile/edit/index.styles.tsx
@@ -1,30 +1,0 @@
-import { css } from '@emotion/react';
-
-export const wrapper = css`
-  padding: 18px;
-`;
-
-export const profileImageWrapper = css`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin: 32px 0 40px;
-`;
-export const inputWrapper = css`
-  margin: 0;
-  padding: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 32px;
-`;
-export const nickNameWrapper = css`
-  display: flex;
-  align-items: flex-end;
-`;
-
-export const readOnlyTextBox = css`
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-`;
