--- apps/daengle/src/pages/mypage/user-profile/edit/interfaces/index.ts
@@ -1,9 +1,0 @@
-export interface UserProfileInfoEditForm {
-  image: File | null;
-  nickname: string;
-  isAvailableNickname: boolean;
-}
-export interface UserProfileInfoEditFormType {
-  form: UserProfileInfoEditForm;
-  isAvailableNickname: boolean;
-}
